Fixed all warning in gcc - Debug mode

+++ include/dirl/baseDynamicsRegressor.hpp
@@ -49,7 +49,7 @@ class baseDynamicsRegressor : public DynamicRegressorInterface
             NrOfRealLinks_subtree = 0;
             for(int ll=0; ll < (int)linkIndeces2regrCols.size(); ll++ ) { if( linkIndeces2regrCols[ll] != -1 ) { NrOfRealLinks_subtree++; } }
             assert(NrOfRealLinks_subtree >= 0);
-            assert(NrOfRealLinks_subtree <= linkIndeces2regrCols.size());
+            assert(NrOfRealLinks_subtree <= (int)linkIndeces2regrCols.size());
         }
                                                                                                                                              
         virtual ~baseDynamicsRegressor() {};
